Query video from DB
